Update RemoveW10Bloatware.ps1
This script will now give you a warning for each app in your $UninstallPackages list that isn't found either in the installed packages or the provisioned packages. Replaced all instances of Write-Host and Write-Warning with calls to Write-Log. As a result, only those specific messages will be written to the log file. It should be much cleaner now.